imprv

+++ core-service/src/main/java/com/patizone/core_service/messaging/WebSocketEventListener.java
@@ -5,6 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.messaging.SessionConnectedEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
@@ -19,11 +20,15 @@ public class WebSocketEventListener {
   @EventListener
   public void handleWebSocketConnectListener(SessionConnectedEvent event) {
     StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-    sessionRegistryService.registerUser(event.getUser().getName(), headerAccessor.getHost());
+    sessionRegistryService.registerUser(event.getUser().getName(), headerAccessor.getSessionId());
   }
 
   @EventListener
   public void handleWebSocketDisconnectListener(SessionDisconnectEvent event) {
-   sessionRegistryService.removeUser(event.getUser().getName());
+    if(event == null) {
+      return;
+    }
+    sessionRegistryService.removeUser(event.getUser().getName());
+    SecurityContextHolder.clearContext();
   }
 }
